Move ColumnInput component to a separate file
