+++ frontend/src/metabase/query_builder/components/expressions/CombineColumns/ColumnAndSeparatorRow/ColumnInput.module.css
@@ -1,0 +1,20 @@
+.button {
+  justify-content: space-between !important;
+  font-weight: normal;
+}
+
+button.root {
+  &:hover {
+    background: none;
+  }
+
+  &:hover .button {
+    color: var(--color-text-dark);
+  }
+
+  &.open,
+  &:focus-visible {
+    border-color: var(--color-brand);
+    outline: none;
+  }
+}
